Skip verifying the peer name
On the production EPP, the common name on the SSL certificate is
epp.nominet.org.uk, you might find that PHP has trouble connecting as it
is expecting it to be ror-epp.nominet.org.uk. There would probably be a
security risk in doing so for a long running project, but for next week,
I think you're safe.