Intermittent fault found when calculating the sleep time, if the sleep time was exactly on a 0 micro second it wasn't appending zeros to fill the time, example of a proper sleep time:
48139267 = 48.139267

However if you happen to hit the calculation on a round number ending in zero, the number was not adding the zero to the end (example below from testing):

4807158   4.807158

which should be:

48071580  48.07158

Because of the missing zero, the timings as you can see are massively different, so of a max will be 5 seconds odd, then you have the extra time doing the EPP creates, and reading before the next log (around 15 seconds most likely).

What happens now is, if its a round number (or any length less than eight it will append the zeros on the end):

Example from testing:

Sleep time (old code):

470553

New Code:

47055300

So will now sleep exactly as it should.

Also added an error catch on no domain file, and added settings for multi create (only tested on a small amount of creates).